Restructured old code into correct format
Also:
- Added .sln file to solution
