+++ Chirp.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{26B2060C-E39C-422E-AAF3-3BFDC937DE65}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.CLI", "src\Chirp.CLI\Chirp.CLI.csproj", "{1932E58B-CC62-4197-96B7-9BDB721A3FCE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleDB", "src\SimpleDB\SimpleDB.csproj", "{985DA78F-657B-43E1-8CBD-EC87913E878E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1932E58B-CC62-4197-96B7-9BDB721A3FCE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1932E58B-CC62-4197-96B7-9BDB721A3FCE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1932E58B-CC62-4197-96B7-9BDB721A3FCE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1932E58B-CC62-4197-96B7-9BDB721A3FCE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{985DA78F-657B-43E1-8CBD-EC87913E878E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{985DA78F-657B-43E1-8CBD-EC87913E878E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{985DA78F-657B-43E1-8CBD-EC87913E878E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{985DA78F-657B-43E1-8CBD-EC87913E878E}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{1932E58B-CC62-4197-96B7-9BDB721A3FCE} = {26B2060C-E39C-422E-AAF3-3BFDC937DE65}
+		{985DA78F-657B-43E1-8CBD-EC87913E878E} = {26B2060C-E39C-422E-AAF3-3BFDC937DE65}
+	EndGlobalSection
+EndGlobal
